first pass at moving cv and class handling into metaclasses
